+++ tempusludicus/Core/Src/pit.c
@@ -8,15 +8,15 @@
 
 #include "pit.h"
 
-volatile uint32_t unix_timestamp = 0;
+volatile uint64_t unix_timestamp = 0;
 // volatile bool pit_interrupt_flag = false;
 
-uint32_t get_unix_timestamp(void)
+uint64_t get_unix_timestamp(void)
 {
 	return unix_timestamp;
 }
 
-void set_unix_timestamp(uint32_t timestamp)
+void set_unix_timestamp(uint64_t timestamp)
 {
 	unix_timestamp = timestamp;
 }
